# docs/swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/locations/{id}": {
      "patch": {
        "operationId": "LocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      }
    },
...
    "/empty-locations/{id}": {
      "get": {
        "description": "Retrieves a single empty location by its unique ID.",
        "operationId": "EmptyLocationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The empty location with the specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location_name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "priority": {
                      "type": "number"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "location_id": {
                      "type": "string"
                    },
                    "is_active": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get an empty location by ID",
        "tags": [
          "EmptyLocations"
        ]
      },
      "patch": {
        "description": "Updates the details of an existing empty location by its ID.",
        "operationId": "EmptyLocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "location_name": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated empty location.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location_name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "priority": {
                      "type": "number"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "location_id": {
                      "type": "string"
                    },
                    "is_active": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Update an empty location",
        "tags": [
          "EmptyLocations"
        ]
      }
    },
    "/empty-locations/update-allocation/{allocation_type}": {
      "patch": {
        "description": "Updates the allocation strategy for empty locations to either ROUND_ROBIN or MANUAL.",
        "operationId": "EmptyLocationsController_updateAllocation",
        "parameters": [
          {
            "name": "allocation_type",
            "required": true,
            "in": "path",
            "description": "Type of allocation strategy to apply",
            "schema": {
              "enum": [
                "ROUND_ROBIN",
                "MANUAL"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The allocation strategy has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "operation_type": {
                      "type": "string",
                      "enum": [
                        "FLOWOPS"
                      ]
                    },
                    "value": {
                      "type": "object",
                      "properties": {
                        "EMPTY_LOCATION": {
                          "type": "string",
                          "enum": [
                            "MANUAL",
                            "ROUND_ROBIN"
                          ]
                        },
                        "AUTO_START": {
                          "type": "integer"
                        },
                        "TASK_ALLOCATION_STRATEGY": {
                          "type": "string",
                          "enum": [
                            "LOAD_BALANCE"
                          ]
                        }
                      },
                      "required": [
                        "EMPTY_LOCATION",
                        "AUTO_START",
                        "TASK_ALLOCATION_STRATEGY"
                      ]
                    }
                  },
                  "required": [
                    "created_at",
                    "updated_at",
                    "id",
                    "operation_type",
                    "value"
                  ]
                }
              }
            }
          }
        },
        "summary": "Update empty location allocation strategy",
        "tags": [
          "EmptyLocations"
        ]
      }
    },
    "/empty-locations/{id}/active-robot": {
      "get": {
        "description": "Retrieves the active robot assigned to a specific empty location by its ID.",
        "operationId": "EmptyLocationsController_getActiveRobot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the empty location",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The active robot at the specified empty location. If no active robot, only robot_id (null) is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "robot_id": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string"
                    },
                    "completed_time": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  },
                  "required": [
                    "robot_id",
                    "source",
                    "status",
                    "completed_time"
                  ]
                }
              }
            }
          }
        },
        "summary": "Get active robot at empty location",
        "tags": [
          "EmptyLocations"
        ]
      }
    },
...
  },
  "info": {
    "title": "ARAPL RaaS GTP API",
    "description": "\n      Goods To Person (GTP) API for ARAPL RaaS (Robotics as a Service)\n\n      This API provides comprehensive GTP capabilities including:\n      - Order management and file upload processing\n      - Moveops location management\n      - Station management and Pallet Slot assignments\n      - Inventory tracking and bulk operations\n      \n      All endpoints include proper validation, error handling, and support file uploads where applicable.\n    ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "API Key for authentication"
      }
    },
    "schemas": {
...
      "UpdateLocationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}